Revert "Hotel Java final step"
