+++ src/files/all_shops.txt
