--- src/files/checkIn.txt
@@ -1,50 +1,0 @@
-Check In Details
-saldjf
-342424
-najk
-jnsfslf
-sdflsf;jl
-2023/04/29
-1000
-Not Booked
-
-Check In Details
-asdfk;nl;slnf
-132424
-fsdfjk
-sdflk;dfs
-sdfkjkd
-2023/04/29
-1000
-Not Booked
-
-Check In Details
-asdfk;nl;slnf
-132424
-fsdfjk
-sdflk;dfs
-sdfkjkd
-2023/04/29
-1000
-Not Booked
-
-Check In Details
-new
-23423
-sadff
-grge
-rgegr
-2023/04/29
-1000
-Not Booked
-
-Check In Details
-sadfsdf
-2342
-sffsdf
-sfasf
-sdfsf
-2023/04/29
-1000
-Not Booked
-
