+++ src/files/rented_shops.txt
