+++ src/myInterface/tableDataReader.java
@@ -1,0 +1,7 @@
+package myInterface;
+
+import javax.swing.table.DefaultTableModel;
+
+public interface tableDataReader {
+	public void readData(DefaultTableModel model);
+}
